requirements.txt
flask==3.1.0
pymysql==1.1.1
python-dotenv==1.0.1
pymysql[rsa]==1.1.1
PyJWT==2.6.0
flask-cors==3.0.10
flask-restx==1.3.0